sneakermerge,module rename
