+++ test/javasource/app/proxies/microflows/Microflows.java
@@ -1,0 +1,67 @@
+// This file was generated by Mendix Modeler 7.23.
+//
+// WARNING: Code you write here will be lost the next time you deploy the project.
+
+package app.proxies.microflows;
+
+import java.util.HashMap;
+import java.util.Map;
+import com.mendix.core.Core;
+import com.mendix.core.CoreException;
+import com.mendix.systemwideinterfaces.MendixRuntimeException;
+import com.mendix.systemwideinterfaces.core.IContext;
+import com.mendix.systemwideinterfaces.core.IMendixObject;
+
+public class Microflows
+{
+	// These are the microflows for the App module
+	public static scenegraph.proxies.Scene ds_scene_warehouse(IContext context)
+	{
+		try
+		{
+			Map<java.lang.String, Object> params = new HashMap<java.lang.String, Object>();
+			IMendixObject result = (IMendixObject)Core.execute(context, "App.ds_scene_warehouse", params);
+			return result == null ? null : scenegraph.proxies.Scene.initialize(context, result);
+		}
+		catch (CoreException e)
+		{
+			throw new MendixRuntimeException(e);
+		}
+	}
+	public static void sub_pallet_add(IContext context, java.lang.Long _palletWidth, java.lang.Long _palletHeight, java.lang.Long _palletDepth, java.lang.Long _palletX, java.lang.Long _palletY, java.lang.Long _palletZ, java.lang.Long _i, java.util.List<scenegraph.proxies.Primitive> _arr_primitive, scenegraph.proxies.Scene _obj_scene, java.util.List<scenegraph.proxies.Node> _arr_node)
+	{
+		try
+		{
+			Map<java.lang.String, Object> params = new HashMap<java.lang.String, Object>();
+			params.put("PalletWidth", _palletWidth);
+			params.put("PalletHeight", _palletHeight);
+			params.put("PalletDepth", _palletDepth);
+			params.put("PalletX", _palletX);
+			params.put("PalletY", _palletY);
+			params.put("PalletZ", _palletZ);
+			params.put("i", _i);
+			java.util.ArrayList<IMendixObject> listparam_arr_primitive = null;
+			if (_arr_primitive != null)
+			{
+				listparam_arr_primitive = new java.util.ArrayList<IMendixObject>();
+				for (scenegraph.proxies.Primitive obj : _arr_primitive)
+					listparam_arr_primitive.add(obj.getMendixObject());
+			}
+			params.put("arr_primitive", listparam_arr_primitive);
+			params.put("obj_scene", _obj_scene == null ? null : _obj_scene.getMendixObject());
+			java.util.ArrayList<IMendixObject> listparam_arr_node = null;
+			if (_arr_node != null)
+			{
+				listparam_arr_node = new java.util.ArrayList<IMendixObject>();
+				for (scenegraph.proxies.Node obj : _arr_node)
+					listparam_arr_node.add(obj.getMendixObject());
+			}
+			params.put("arr_node", listparam_arr_node);
+			Core.execute(context, "App.sub_pallet_add", params);
+		}
+		catch (CoreException e)
+		{
+			throw new MendixRuntimeException(e);
+		}
+	}
+}
