--- test/javasource/main/actions/ja_imendixobject2json.java
@@ -1,66 +1,0 @@
-// This file was generated by Mendix Modeler.
-//
-// WARNING: Only the following code will be retained when actions are regenerated:
-// - the import list
-// - the code between BEGIN USER CODE and END USER CODE
-// - the code between BEGIN EXTRA CODE and END EXTRA CODE
-// Other code you write will be lost the next time you deploy the project.
-// Special characters, e.g., é, ö, à, etc. are supported in comments.
-
-package main.actions;
-
-import com.mendix.systemwideinterfaces.core.IContext;
-import com.mendix.webui.CustomJavaAction;
-import com.mendix.systemwideinterfaces.core.IMendixObject;
-
-/**
- * serializes com.mendix.systemwideinterfaces.core.IMendixObject to JSON string
- * 
- * status:in progress
- */
-public class ja_imendixobject2json extends CustomJavaAction<java.lang.String>
-{
-	private IMendixObject Parameter;
-
-	public ja_imendixobject2json(IContext context, IMendixObject Parameter)
-	{
-		super(context);
-		this.Parameter = Parameter;
-	}
-
-	@java.lang.Override
-	public java.lang.String executeAction() throws Exception
-	{
-		// BEGIN USER CODE
-		java.lang.String ret="";
-		if(Parameter!=null){
-			try{
-				ret+=Parameter.getMetaObject().getName();
-				ret+="\n";
-				for(com.mendix.systemwideinterfaces.core.meta.IMetaPrimitive mp:Parameter.getMetaObject().getMetaPrimitives()) {
-					ret+=mp.getName();
-					ret+=":";
-					ret+=mp.getName().toString();
-					ret+="\n";
-				}
-			}catch(Exception e){
-				com.mendix.core.Core.getLogger(this.toString()).error(e.toString());
-			}		   
-		}else{
-		}
-		return ret;
-		// END USER CODE
-	}
-
-	/**
-	 * Returns a string representation of this action
-	 */
-	@java.lang.Override
-	public java.lang.String toString()
-	{
-		return "ja_imendixobject2json";
-	}
-
-	// BEGIN EXTRA CODE
-	// END EXTRA CODE
-}
