--- test/javasource/main/proxies/Camera.java
@@ -1,436 +1,0 @@
-// This file was generated by Mendix Modeler.
-//
-// WARNING: Code you write here will be lost the next time you deploy the project.
-
-package main.proxies;
-
-public class Camera
-{
-	private final com.mendix.systemwideinterfaces.core.IMendixObject cameraMendixObject;
-
-	private final com.mendix.systemwideinterfaces.core.IContext context;
-
-	/**
-	 * Internal name of this entity
-	 */
-	public static final java.lang.String entityName = "Main.Camera";
-
-	/**
-	 * Enum describing members of this entity
-	 */
-	public enum MemberNames
-	{
-		x("x"),
-		y("y"),
-		z("z"),
-		tx("tx"),
-		ty("ty"),
-		tz("tz"),
-		Scene_Camera("Main.Scene_Camera");
-
-		private java.lang.String metaName;
-
-		MemberNames(java.lang.String s)
-		{
-			metaName = s;
-		}
-
-		@java.lang.Override
-		public java.lang.String toString()
-		{
-			return metaName;
-		}
-	}
-
-	public Camera(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		this(context, com.mendix.core.Core.instantiate(context, "Main.Camera"));
-	}
-
-	protected Camera(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject cameraMendixObject)
-	{
-		if (cameraMendixObject == null)
-			throw new java.lang.IllegalArgumentException("The given object cannot be null.");
-		if (!com.mendix.core.Core.isSubClassOf("Main.Camera", cameraMendixObject.getType()))
-			throw new java.lang.IllegalArgumentException("The given object is not a Main.Camera");
-
-		this.cameraMendixObject = cameraMendixObject;
-		this.context = context;
-	}
-
-	/**
-	 * @deprecated Use 'Camera.load(IContext, IMendixIdentifier)' instead.
-	 */
-	@java.lang.Deprecated
-	public static main.proxies.Camera initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		return main.proxies.Camera.load(context, mendixIdentifier);
-	}
-
-	/**
-	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
-	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
-	 */
-	public static main.proxies.Camera initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
-	{
-		return new main.proxies.Camera(context, mendixObject);
-	}
-
-	public static main.proxies.Camera load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
-		return main.proxies.Camera.initialize(context, mendixObject);
-	}
-
-	public static java.util.List<main.proxies.Camera> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
-	{
-		java.util.List<main.proxies.Camera> result = new java.util.ArrayList<main.proxies.Camera>();
-		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//Main.Camera" + xpathConstraint))
-			result.add(main.proxies.Camera.initialize(context, obj));
-		return result;
-	}
-
-	/**
-	 * Commit the changes made on this proxy object.
-	 */
-	public final void commit() throws com.mendix.core.CoreException
-	{
-		com.mendix.core.Core.commit(context, getMendixObject());
-	}
-
-	/**
-	 * Commit the changes made on this proxy object using the specified context.
-	 */
-	public final void commit(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
-	{
-		com.mendix.core.Core.commit(context, getMendixObject());
-	}
-
-	/**
-	 * Delete the object.
-	 */
-	public final void delete()
-	{
-		com.mendix.core.Core.delete(context, getMendixObject());
-	}
-
-	/**
-	 * Delete the object using the specified context.
-	 */
-	public final void delete(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		com.mendix.core.Core.delete(context, getMendixObject());
-	}
-	/**
-	 * @return value of x
-	 */
-	public final java.lang.Double getx()
-	{
-		return getx(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of x
-	 */
-	public final java.lang.Double getx(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.x.toString());
-	}
-
-	/**
-	 * Set value of x
-	 * @param x
-	 */
-	public final void setx(java.lang.Double x)
-	{
-		setx(getContext(), x);
-	}
-
-	/**
-	 * Set value of x
-	 * @param context
-	 * @param x
-	 */
-	public final void setx(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double x)
-	{
-		getMendixObject().setValue(context, MemberNames.x.toString(), x);
-	}
-
-	/**
-	 * @return value of y
-	 */
-	public final java.lang.Double gety()
-	{
-		return gety(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of y
-	 */
-	public final java.lang.Double gety(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.y.toString());
-	}
-
-	/**
-	 * Set value of y
-	 * @param y
-	 */
-	public final void sety(java.lang.Double y)
-	{
-		sety(getContext(), y);
-	}
-
-	/**
-	 * Set value of y
-	 * @param context
-	 * @param y
-	 */
-	public final void sety(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double y)
-	{
-		getMendixObject().setValue(context, MemberNames.y.toString(), y);
-	}
-
-	/**
-	 * @return value of z
-	 */
-	public final java.lang.Double getz()
-	{
-		return getz(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of z
-	 */
-	public final java.lang.Double getz(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.z.toString());
-	}
-
-	/**
-	 * Set value of z
-	 * @param z
-	 */
-	public final void setz(java.lang.Double z)
-	{
-		setz(getContext(), z);
-	}
-
-	/**
-	 * Set value of z
-	 * @param context
-	 * @param z
-	 */
-	public final void setz(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double z)
-	{
-		getMendixObject().setValue(context, MemberNames.z.toString(), z);
-	}
-
-	/**
-	 * @return value of tx
-	 */
-	public final java.lang.Double gettx()
-	{
-		return gettx(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of tx
-	 */
-	public final java.lang.Double gettx(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.tx.toString());
-	}
-
-	/**
-	 * Set value of tx
-	 * @param tx
-	 */
-	public final void settx(java.lang.Double tx)
-	{
-		settx(getContext(), tx);
-	}
-
-	/**
-	 * Set value of tx
-	 * @param context
-	 * @param tx
-	 */
-	public final void settx(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double tx)
-	{
-		getMendixObject().setValue(context, MemberNames.tx.toString(), tx);
-	}
-
-	/**
-	 * @return value of ty
-	 */
-	public final java.lang.Double getty()
-	{
-		return getty(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of ty
-	 */
-	public final java.lang.Double getty(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.ty.toString());
-	}
-
-	/**
-	 * Set value of ty
-	 * @param ty
-	 */
-	public final void setty(java.lang.Double ty)
-	{
-		setty(getContext(), ty);
-	}
-
-	/**
-	 * Set value of ty
-	 * @param context
-	 * @param ty
-	 */
-	public final void setty(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double ty)
-	{
-		getMendixObject().setValue(context, MemberNames.ty.toString(), ty);
-	}
-
-	/**
-	 * @return value of tz
-	 */
-	public final java.lang.Double gettz()
-	{
-		return gettz(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of tz
-	 */
-	public final java.lang.Double gettz(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.tz.toString());
-	}
-
-	/**
-	 * Set value of tz
-	 * @param tz
-	 */
-	public final void settz(java.lang.Double tz)
-	{
-		settz(getContext(), tz);
-	}
-
-	/**
-	 * Set value of tz
-	 * @param context
-	 * @param tz
-	 */
-	public final void settz(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double tz)
-	{
-		getMendixObject().setValue(context, MemberNames.tz.toString(), tz);
-	}
-
-	/**
-	 * @return value of Scene_Camera
-	 */
-	public final main.proxies.Scene getScene_Camera() throws com.mendix.core.CoreException
-	{
-		return getScene_Camera(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of Scene_Camera
-	 */
-	public final main.proxies.Scene getScene_Camera(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
-	{
-		main.proxies.Scene result = null;
-		com.mendix.systemwideinterfaces.core.IMendixIdentifier identifier = getMendixObject().getValue(context, MemberNames.Scene_Camera.toString());
-		if (identifier != null)
-			result = main.proxies.Scene.load(context, identifier);
-		return result;
-	}
-
-	/**
-	 * Set value of Scene_Camera
-	 * @param scene_camera
-	 */
-	public final void setScene_Camera(main.proxies.Scene scene_camera)
-	{
-		setScene_Camera(getContext(), scene_camera);
-	}
-
-	/**
-	 * Set value of Scene_Camera
-	 * @param context
-	 * @param scene_camera
-	 */
-	public final void setScene_Camera(com.mendix.systemwideinterfaces.core.IContext context, main.proxies.Scene scene_camera)
-	{
-		if (scene_camera == null)
-			getMendixObject().setValue(context, MemberNames.Scene_Camera.toString(), null);
-		else
-			getMendixObject().setValue(context, MemberNames.Scene_Camera.toString(), scene_camera.getMendixObject().getId());
-	}
-
-	/**
-	 * @return the IMendixObject instance of this proxy for use in the Core interface.
-	 */
-	public final com.mendix.systemwideinterfaces.core.IMendixObject getMendixObject()
-	{
-		return cameraMendixObject;
-	}
-
-	/**
-	 * @return the IContext instance of this proxy, or null if no IContext instance was specified at initialization.
-	 */
-	public final com.mendix.systemwideinterfaces.core.IContext getContext()
-	{
-		return context;
-	}
-
-	@java.lang.Override
-	public boolean equals(Object obj)
-	{
-		if (obj == this)
-			return true;
-
-		if (obj != null && getClass().equals(obj.getClass()))
-		{
-			final main.proxies.Camera that = (main.proxies.Camera) obj;
-			return getMendixObject().equals(that.getMendixObject());
-		}
-		return false;
-	}
-
-	@java.lang.Override
-	public int hashCode()
-	{
-		return getMendixObject().hashCode();
-	}
-
-	/**
-	 * @return String name of this class
-	 */
-	public static java.lang.String getType()
-	{
-		return "Main.Camera";
-	}
-
-	/**
-	 * @return String GUID from this object, format: ID_0000000000
-	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
-	 */
-	@java.lang.Deprecated
-	public java.lang.String getGUID()
-	{
-		return "ID_" + getMendixObject().getId().toLong();
-	}
-}
