--- test/javasource/main/proxies/Image.java
@@ -1,123 +1,0 @@
-// This file was generated by Mendix Modeler.
-//
-// WARNING: Code you write here will be lost the next time you deploy the project.
-
-package main.proxies;
-
-public class Image extends system.proxies.Image
-{
-	/**
-	 * Internal name of this entity
-	 */
-	public static final java.lang.String entityName = "Main.Image";
-
-	/**
-	 * Enum describing members of this entity
-	 */
-	public enum MemberNames
-	{
-		PublicThumbnailPath("PublicThumbnailPath"),
-		FileID("FileID"),
-		Name("Name"),
-		DeleteAfterDownload("DeleteAfterDownload"),
-		Contents("Contents"),
-		HasContents("HasContents"),
-		Size("Size");
-
-		private java.lang.String metaName;
-
-		MemberNames(java.lang.String s)
-		{
-			metaName = s;
-		}
-
-		@java.lang.Override
-		public java.lang.String toString()
-		{
-			return metaName;
-		}
-	}
-
-	public Image(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		this(context, com.mendix.core.Core.instantiate(context, "Main.Image"));
-	}
-
-	protected Image(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject imageMendixObject)
-	{
-		super(context, imageMendixObject);
-		if (!com.mendix.core.Core.isSubClassOf("Main.Image", imageMendixObject.getType()))
-			throw new java.lang.IllegalArgumentException("The given object is not a Main.Image");
-	}
-
-	/**
-	 * @deprecated Use 'Image.load(IContext, IMendixIdentifier)' instead.
-	 */
-	@java.lang.Deprecated
-	public static main.proxies.Image initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		return main.proxies.Image.load(context, mendixIdentifier);
-	}
-
-	/**
-	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
-	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
-	 */
-	public static main.proxies.Image initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
-	{
-		return new main.proxies.Image(context, mendixObject);
-	}
-
-	public static main.proxies.Image load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
-		return main.proxies.Image.initialize(context, mendixObject);
-	}
-
-	public static java.util.List<main.proxies.Image> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
-	{
-		java.util.List<main.proxies.Image> result = new java.util.ArrayList<main.proxies.Image>();
-		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//Main.Image" + xpathConstraint))
-			result.add(main.proxies.Image.initialize(context, obj));
-		return result;
-	}
-
-	@java.lang.Override
-	public boolean equals(Object obj)
-	{
-		if (obj == this)
-			return true;
-
-		if (obj != null && getClass().equals(obj.getClass()))
-		{
-			final main.proxies.Image that = (main.proxies.Image) obj;
-			return getMendixObject().equals(that.getMendixObject());
-		}
-		return false;
-	}
-
-	@java.lang.Override
-	public int hashCode()
-	{
-		return getMendixObject().hashCode();
-	}
-
-	/**
-	 * @return String name of this class
-	 */
-	public static java.lang.String getType()
-	{
-		return "Main.Image";
-	}
-
-	/**
-	 * @return String GUID from this object, format: ID_0000000000
-	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
-	 */
-	@java.lang.Override
-	@java.lang.Deprecated
-	public java.lang.String getGUID()
-	{
-		return "ID_" + getMendixObject().getId().toLong();
-	}
-}
