--- test/javasource/main/proxies/Line.java
@@ -1,239 +1,0 @@
-// This file was generated by Mendix Modeler.
-//
-// WARNING: Code you write here will be lost the next time you deploy the project.
-
-package main.proxies;
-
-public class Line extends main.proxies.Primitive
-{
-	/**
-	 * Internal name of this entity
-	 */
-	public static final java.lang.String entityName = "Main.Line";
-
-	/**
-	 * Enum describing members of this entity
-	 */
-	public enum MemberNames
-	{
-		x1("x1"),
-		y1("y1"),
-		z1("z1"),
-		Name("Name"),
-		x("x"),
-		y("y"),
-		z("z"),
-		rotx("rotx"),
-		roty("roty"),
-		rotz("rotz"),
-		visible("visible"),
-		color("color"),
-		Primitive_Node("Main.Primitive_Node"),
-		Selected("Main.Selected"),
-		Texture("Main.Texture");
-
-		private java.lang.String metaName;
-
-		MemberNames(java.lang.String s)
-		{
-			metaName = s;
-		}
-
-		@java.lang.Override
-		public java.lang.String toString()
-		{
-			return metaName;
-		}
-	}
-
-	public Line(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		this(context, com.mendix.core.Core.instantiate(context, "Main.Line"));
-	}
-
-	protected Line(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject lineMendixObject)
-	{
-		super(context, lineMendixObject);
-		if (!com.mendix.core.Core.isSubClassOf("Main.Line", lineMendixObject.getType()))
-			throw new java.lang.IllegalArgumentException("The given object is not a Main.Line");
-	}
-
-	/**
-	 * @deprecated Use 'Line.load(IContext, IMendixIdentifier)' instead.
-	 */
-	@java.lang.Deprecated
-	public static main.proxies.Line initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		return main.proxies.Line.load(context, mendixIdentifier);
-	}
-
-	/**
-	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
-	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
-	 */
-	public static main.proxies.Line initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
-	{
-		return new main.proxies.Line(context, mendixObject);
-	}
-
-	public static main.proxies.Line load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
-		return main.proxies.Line.initialize(context, mendixObject);
-	}
-
-	public static java.util.List<main.proxies.Line> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
-	{
-		java.util.List<main.proxies.Line> result = new java.util.ArrayList<main.proxies.Line>();
-		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//Main.Line" + xpathConstraint))
-			result.add(main.proxies.Line.initialize(context, obj));
-		return result;
-	}
-
-	/**
-	 * @return value of x1
-	 */
-	public final java.lang.Double getx1()
-	{
-		return getx1(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of x1
-	 */
-	public final java.lang.Double getx1(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.x1.toString());
-	}
-
-	/**
-	 * Set value of x1
-	 * @param x1
-	 */
-	public final void setx1(java.lang.Double x1)
-	{
-		setx1(getContext(), x1);
-	}
-
-	/**
-	 * Set value of x1
-	 * @param context
-	 * @param x1
-	 */
-	public final void setx1(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double x1)
-	{
-		getMendixObject().setValue(context, MemberNames.x1.toString(), x1);
-	}
-
-	/**
-	 * @return value of y1
-	 */
-	public final java.lang.Double gety1()
-	{
-		return gety1(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of y1
-	 */
-	public final java.lang.Double gety1(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.y1.toString());
-	}
-
-	/**
-	 * Set value of y1
-	 * @param y1
-	 */
-	public final void sety1(java.lang.Double y1)
-	{
-		sety1(getContext(), y1);
-	}
-
-	/**
-	 * Set value of y1
-	 * @param context
-	 * @param y1
-	 */
-	public final void sety1(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double y1)
-	{
-		getMendixObject().setValue(context, MemberNames.y1.toString(), y1);
-	}
-
-	/**
-	 * @return value of z1
-	 */
-	public final java.lang.Double getz1()
-	{
-		return getz1(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of z1
-	 */
-	public final java.lang.Double getz1(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.z1.toString());
-	}
-
-	/**
-	 * Set value of z1
-	 * @param z1
-	 */
-	public final void setz1(java.lang.Double z1)
-	{
-		setz1(getContext(), z1);
-	}
-
-	/**
-	 * Set value of z1
-	 * @param context
-	 * @param z1
-	 */
-	public final void setz1(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double z1)
-	{
-		getMendixObject().setValue(context, MemberNames.z1.toString(), z1);
-	}
-
-	@java.lang.Override
-	public boolean equals(Object obj)
-	{
-		if (obj == this)
-			return true;
-
-		if (obj != null && getClass().equals(obj.getClass()))
-		{
-			final main.proxies.Line that = (main.proxies.Line) obj;
-			return getMendixObject().equals(that.getMendixObject());
-		}
-		return false;
-	}
-
-	@java.lang.Override
-	public int hashCode()
-	{
-		return getMendixObject().hashCode();
-	}
-
-	/**
-	 * @return String name of this class
-	 */
-	public static java.lang.String getType()
-	{
-		return "Main.Line";
-	}
-
-	/**
-	 * @return String GUID from this object, format: ID_0000000000
-	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
-	 */
-	@java.lang.Override
-	@java.lang.Deprecated
-	public java.lang.String getGUID()
-	{
-		return "ID_" + getMendixObject().getId().toLong();
-	}
-}
