--- test/javasource/main/proxies/Model.java
@@ -1,172 +1,0 @@
-// This file was generated by Mendix Modeler.
-//
-// WARNING: Code you write here will be lost the next time you deploy the project.
-
-package main.proxies;
-
-public class Model extends main.proxies.Primitive
-{
-	/**
-	 * Internal name of this entity
-	 */
-	public static final java.lang.String entityName = "Main.Model";
-
-	/**
-	 * Enum describing members of this entity
-	 */
-	public enum MemberNames
-	{
-		Name("Name"),
-		x("x"),
-		y("y"),
-		z("z"),
-		rotx("rotx"),
-		roty("roty"),
-		rotz("rotz"),
-		visible("visible"),
-		color("color"),
-		Model_ModelFile("Main.Model_ModelFile"),
-		Primitive_Node("Main.Primitive_Node"),
-		Selected("Main.Selected"),
-		Texture("Main.Texture");
-
-		private java.lang.String metaName;
-
-		MemberNames(java.lang.String s)
-		{
-			metaName = s;
-		}
-
-		@java.lang.Override
-		public java.lang.String toString()
-		{
-			return metaName;
-		}
-	}
-
-	public Model(com.mendix.systemwideinterfaces.core.IContext context)
-	{
-		this(context, com.mendix.core.Core.instantiate(context, "Main.Model"));
-	}
-
-	protected Model(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject modelMendixObject)
-	{
-		super(context, modelMendixObject);
-		if (!com.mendix.core.Core.isSubClassOf("Main.Model", modelMendixObject.getType()))
-			throw new java.lang.IllegalArgumentException("The given object is not a Main.Model");
-	}
-
-	/**
-	 * @deprecated Use 'Model.load(IContext, IMendixIdentifier)' instead.
-	 */
-	@java.lang.Deprecated
-	public static main.proxies.Model initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		return main.proxies.Model.load(context, mendixIdentifier);
-	}
-
-	/**
-	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
-	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
-	 */
-	public static main.proxies.Model initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
-	{
-		return new main.proxies.Model(context, mendixObject);
-	}
-
-	public static main.proxies.Model load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
-	{
-		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
-		return main.proxies.Model.initialize(context, mendixObject);
-	}
-
-	public static java.util.List<main.proxies.Model> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
-	{
-		java.util.List<main.proxies.Model> result = new java.util.ArrayList<main.proxies.Model>();
-		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//Main.Model" + xpathConstraint))
-			result.add(main.proxies.Model.initialize(context, obj));
-		return result;
-	}
-
-	/**
-	 * @return value of Model_ModelFile
-	 */
-	public final main.proxies.ModelFile getModel_ModelFile() throws com.mendix.core.CoreException
-	{
-		return getModel_ModelFile(getContext());
-	}
-
-	/**
-	 * @param context
-	 * @return value of Model_ModelFile
-	 */
-	public final main.proxies.ModelFile getModel_ModelFile(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
-	{
-		main.proxies.ModelFile result = null;
-		com.mendix.systemwideinterfaces.core.IMendixIdentifier identifier = getMendixObject().getValue(context, MemberNames.Model_ModelFile.toString());
-		if (identifier != null)
-			result = main.proxies.ModelFile.load(context, identifier);
-		return result;
-	}
-
-	/**
-	 * Set value of Model_ModelFile
-	 * @param model_modelfile
-	 */
-	public final void setModel_ModelFile(main.proxies.ModelFile model_modelfile)
-	{
-		setModel_ModelFile(getContext(), model_modelfile);
-	}
-
-	/**
-	 * Set value of Model_ModelFile
-	 * @param context
-	 * @param model_modelfile
-	 */
-	public final void setModel_ModelFile(com.mendix.systemwideinterfaces.core.IContext context, main.proxies.ModelFile model_modelfile)
-	{
-		if (model_modelfile == null)
-			getMendixObject().setValue(context, MemberNames.Model_ModelFile.toString(), null);
-		else
-			getMendixObject().setValue(context, MemberNames.Model_ModelFile.toString(), model_modelfile.getMendixObject().getId());
-	}
-
-	@java.lang.Override
-	public boolean equals(Object obj)
-	{
-		if (obj == this)
-			return true;
-
-		if (obj != null && getClass().equals(obj.getClass()))
-		{
-			final main.proxies.Model that = (main.proxies.Model) obj;
-			return getMendixObject().equals(that.getMendixObject());
-		}
-		return false;
-	}
-
-	@java.lang.Override
-	public int hashCode()
-	{
-		return getMendixObject().hashCode();
-	}
-
-	/**
-	 * @return String name of this class
-	 */
-	public static java.lang.String getType()
-	{
-		return "Main.Model";
-	}
-
-	/**
-	 * @return String GUID from this object, format: ID_0000000000
-	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
-	 */
-	@java.lang.Override
-	@java.lang.Deprecated
-	public java.lang.String getGUID()
-	{
-		return "ID_" + getMendixObject().getId().toLong();
-	}
-}
