--- test/javasource/main/proxies/constants/Constants.java
@@ -1,17 +1,0 @@
-// This file was generated by Mendix Modeler 7.23.
-//
-// WARNING: Code you write here will be lost the next time you deploy the project.
-
-package main.proxies.constants;
-
-import com.mendix.core.Core;
-
-public class Constants
-{
-	// These are the constants for the Main module
-
-	public static java.lang.Double getPI()
-	{
-		return (java.lang.Double)Core.getConfiguration().getConstantValue("Main.PI");
-	}
-}
