+++ test/javasource/scenegraph/proxies/FileDocument.java
@@ -1,0 +1,126 @@
+// This file was generated by Mendix Modeler.
+//
+// WARNING: Code you write here will be lost the next time you deploy the project.
+
+package scenegraph.proxies;
+
+/**
+ * uses:
+ *     import zip
+ */
+public class FileDocument extends system.proxies.FileDocument
+{
+	/**
+	 * Internal name of this entity
+	 */
+	public static final java.lang.String entityName = "SceneGraph.FileDocument";
+
+	/**
+	 * Enum describing members of this entity
+	 */
+	public enum MemberNames
+	{
+		FileID("FileID"),
+		Name("Name"),
+		DeleteAfterDownload("DeleteAfterDownload"),
+		Contents("Contents"),
+		HasContents("HasContents"),
+		Size("Size");
+
+		private java.lang.String metaName;
+
+		MemberNames(java.lang.String s)
+		{
+			metaName = s;
+		}
+
+		@java.lang.Override
+		public java.lang.String toString()
+		{
+			return metaName;
+		}
+	}
+
+	public FileDocument(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		this(context, com.mendix.core.Core.instantiate(context, "SceneGraph.FileDocument"));
+	}
+
+	protected FileDocument(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject fileDocumentMendixObject)
+	{
+		super(context, fileDocumentMendixObject);
+		if (!com.mendix.core.Core.isSubClassOf("SceneGraph.FileDocument", fileDocumentMendixObject.getType()))
+			throw new java.lang.IllegalArgumentException("The given object is not a SceneGraph.FileDocument");
+	}
+
+	/**
+	 * @deprecated Use 'FileDocument.load(IContext, IMendixIdentifier)' instead.
+	 */
+	@java.lang.Deprecated
+	public static scenegraph.proxies.FileDocument initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		return scenegraph.proxies.FileDocument.load(context, mendixIdentifier);
+	}
+
+	/**
+	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
+	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
+	 */
+	public static scenegraph.proxies.FileDocument initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
+	{
+		return new scenegraph.proxies.FileDocument(context, mendixObject);
+	}
+
+	public static scenegraph.proxies.FileDocument load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
+		return scenegraph.proxies.FileDocument.initialize(context, mendixObject);
+	}
+
+	public static java.util.List<scenegraph.proxies.FileDocument> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
+	{
+		java.util.List<scenegraph.proxies.FileDocument> result = new java.util.ArrayList<scenegraph.proxies.FileDocument>();
+		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//SceneGraph.FileDocument" + xpathConstraint))
+			result.add(scenegraph.proxies.FileDocument.initialize(context, obj));
+		return result;
+	}
+
+	@java.lang.Override
+	public boolean equals(Object obj)
+	{
+		if (obj == this)
+			return true;
+
+		if (obj != null && getClass().equals(obj.getClass()))
+		{
+			final scenegraph.proxies.FileDocument that = (scenegraph.proxies.FileDocument) obj;
+			return getMendixObject().equals(that.getMendixObject());
+		}
+		return false;
+	}
+
+	@java.lang.Override
+	public int hashCode()
+	{
+		return getMendixObject().hashCode();
+	}
+
+	/**
+	 * @return String name of this class
+	 */
+	public static java.lang.String getType()
+	{
+		return "SceneGraph.FileDocument";
+	}
+
+	/**
+	 * @return String GUID from this object, format: ID_0000000000
+	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
+	 */
+	@java.lang.Override
+	@java.lang.Deprecated
+	public java.lang.String getGUID()
+	{
+		return "ID_" + getMendixObject().getId().toLong();
+	}
+}
