+++ test/javasource/scenegraph/proxies/Model.java
@@ -1,0 +1,172 @@
+// This file was generated by Mendix Modeler.
+//
+// WARNING: Code you write here will be lost the next time you deploy the project.
+
+package scenegraph.proxies;
+
+public class Model extends scenegraph.proxies.Primitive
+{
+	/**
+	 * Internal name of this entity
+	 */
+	public static final java.lang.String entityName = "SceneGraph.Model";
+
+	/**
+	 * Enum describing members of this entity
+	 */
+	public enum MemberNames
+	{
+		Name("Name"),
+		x("x"),
+		y("y"),
+		z("z"),
+		rotx("rotx"),
+		roty("roty"),
+		rotz("rotz"),
+		visible("visible"),
+		color("color"),
+		Model_ModelFile("SceneGraph.Model_ModelFile"),
+		Primitive_Node("SceneGraph.Primitive_Node"),
+		Selected("SceneGraph.Selected"),
+		Texture("SceneGraph.Texture");
+
+		private java.lang.String metaName;
+
+		MemberNames(java.lang.String s)
+		{
+			metaName = s;
+		}
+
+		@java.lang.Override
+		public java.lang.String toString()
+		{
+			return metaName;
+		}
+	}
+
+	public Model(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		this(context, com.mendix.core.Core.instantiate(context, "SceneGraph.Model"));
+	}
+
+	protected Model(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject modelMendixObject)
+	{
+		super(context, modelMendixObject);
+		if (!com.mendix.core.Core.isSubClassOf("SceneGraph.Model", modelMendixObject.getType()))
+			throw new java.lang.IllegalArgumentException("The given object is not a SceneGraph.Model");
+	}
+
+	/**
+	 * @deprecated Use 'Model.load(IContext, IMendixIdentifier)' instead.
+	 */
+	@java.lang.Deprecated
+	public static scenegraph.proxies.Model initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		return scenegraph.proxies.Model.load(context, mendixIdentifier);
+	}
+
+	/**
+	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
+	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
+	 */
+	public static scenegraph.proxies.Model initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
+	{
+		return new scenegraph.proxies.Model(context, mendixObject);
+	}
+
+	public static scenegraph.proxies.Model load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
+		return scenegraph.proxies.Model.initialize(context, mendixObject);
+	}
+
+	public static java.util.List<scenegraph.proxies.Model> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
+	{
+		java.util.List<scenegraph.proxies.Model> result = new java.util.ArrayList<scenegraph.proxies.Model>();
+		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//SceneGraph.Model" + xpathConstraint))
+			result.add(scenegraph.proxies.Model.initialize(context, obj));
+		return result;
+	}
+
+	/**
+	 * @return value of Model_ModelFile
+	 */
+	public final scenegraph.proxies.ModelFile getModel_ModelFile() throws com.mendix.core.CoreException
+	{
+		return getModel_ModelFile(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of Model_ModelFile
+	 */
+	public final scenegraph.proxies.ModelFile getModel_ModelFile(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
+	{
+		scenegraph.proxies.ModelFile result = null;
+		com.mendix.systemwideinterfaces.core.IMendixIdentifier identifier = getMendixObject().getValue(context, MemberNames.Model_ModelFile.toString());
+		if (identifier != null)
+			result = scenegraph.proxies.ModelFile.load(context, identifier);
+		return result;
+	}
+
+	/**
+	 * Set value of Model_ModelFile
+	 * @param model_modelfile
+	 */
+	public final void setModel_ModelFile(scenegraph.proxies.ModelFile model_modelfile)
+	{
+		setModel_ModelFile(getContext(), model_modelfile);
+	}
+
+	/**
+	 * Set value of Model_ModelFile
+	 * @param context
+	 * @param model_modelfile
+	 */
+	public final void setModel_ModelFile(com.mendix.systemwideinterfaces.core.IContext context, scenegraph.proxies.ModelFile model_modelfile)
+	{
+		if (model_modelfile == null)
+			getMendixObject().setValue(context, MemberNames.Model_ModelFile.toString(), null);
+		else
+			getMendixObject().setValue(context, MemberNames.Model_ModelFile.toString(), model_modelfile.getMendixObject().getId());
+	}
+
+	@java.lang.Override
+	public boolean equals(Object obj)
+	{
+		if (obj == this)
+			return true;
+
+		if (obj != null && getClass().equals(obj.getClass()))
+		{
+			final scenegraph.proxies.Model that = (scenegraph.proxies.Model) obj;
+			return getMendixObject().equals(that.getMendixObject());
+		}
+		return false;
+	}
+
+	@java.lang.Override
+	public int hashCode()
+	{
+		return getMendixObject().hashCode();
+	}
+
+	/**
+	 * @return String name of this class
+	 */
+	public static java.lang.String getType()
+	{
+		return "SceneGraph.Model";
+	}
+
+	/**
+	 * @return String GUID from this object, format: ID_0000000000
+	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
+	 */
+	@java.lang.Override
+	@java.lang.Deprecated
+	public java.lang.String getGUID()
+	{
+		return "ID_" + getMendixObject().getId().toLong();
+	}
+}
