+++ test/javasource/scenegraph/proxies/Node.java
@@ -1,0 +1,295 @@
+// This file was generated by Mendix Modeler.
+//
+// WARNING: Code you write here will be lost the next time you deploy the project.
+
+package scenegraph.proxies;
+
+public class Node
+{
+	private final com.mendix.systemwideinterfaces.core.IMendixObject nodeMendixObject;
+
+	private final com.mendix.systemwideinterfaces.core.IContext context;
+
+	/**
+	 * Internal name of this entity
+	 */
+	public static final java.lang.String entityName = "SceneGraph.Node";
+
+	/**
+	 * Enum describing members of this entity
+	 */
+	public enum MemberNames
+	{
+		_id("_id"),
+		Node_Scene("SceneGraph.Node_Scene"),
+		Node_Node("SceneGraph.Node_Node");
+
+		private java.lang.String metaName;
+
+		MemberNames(java.lang.String s)
+		{
+			metaName = s;
+		}
+
+		@java.lang.Override
+		public java.lang.String toString()
+		{
+			return metaName;
+		}
+	}
+
+	public Node(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		this(context, com.mendix.core.Core.instantiate(context, "SceneGraph.Node"));
+	}
+
+	protected Node(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject nodeMendixObject)
+	{
+		if (nodeMendixObject == null)
+			throw new java.lang.IllegalArgumentException("The given object cannot be null.");
+		if (!com.mendix.core.Core.isSubClassOf("SceneGraph.Node", nodeMendixObject.getType()))
+			throw new java.lang.IllegalArgumentException("The given object is not a SceneGraph.Node");
+
+		this.nodeMendixObject = nodeMendixObject;
+		this.context = context;
+	}
+
+	/**
+	 * @deprecated Use 'Node.load(IContext, IMendixIdentifier)' instead.
+	 */
+	@java.lang.Deprecated
+	public static scenegraph.proxies.Node initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		return scenegraph.proxies.Node.load(context, mendixIdentifier);
+	}
+
+	/**
+	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
+	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
+	 */
+	public static scenegraph.proxies.Node initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
+	{
+		return new scenegraph.proxies.Node(context, mendixObject);
+	}
+
+	public static scenegraph.proxies.Node load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
+		return scenegraph.proxies.Node.initialize(context, mendixObject);
+	}
+
+	public static java.util.List<scenegraph.proxies.Node> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
+	{
+		java.util.List<scenegraph.proxies.Node> result = new java.util.ArrayList<scenegraph.proxies.Node>();
+		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//SceneGraph.Node" + xpathConstraint))
+			result.add(scenegraph.proxies.Node.initialize(context, obj));
+		return result;
+	}
+
+	/**
+	 * Commit the changes made on this proxy object.
+	 */
+	public final void commit() throws com.mendix.core.CoreException
+	{
+		com.mendix.core.Core.commit(context, getMendixObject());
+	}
+
+	/**
+	 * Commit the changes made on this proxy object using the specified context.
+	 */
+	public final void commit(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
+	{
+		com.mendix.core.Core.commit(context, getMendixObject());
+	}
+
+	/**
+	 * Delete the object.
+	 */
+	public final void delete()
+	{
+		com.mendix.core.Core.delete(context, getMendixObject());
+	}
+
+	/**
+	 * Delete the object using the specified context.
+	 */
+	public final void delete(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		com.mendix.core.Core.delete(context, getMendixObject());
+	}
+	/**
+	 * @return value of _id
+	 */
+	public final java.lang.Long get_id()
+	{
+		return get_id(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of _id
+	 */
+	public final java.lang.Long get_id(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		return (java.lang.Long) getMendixObject().getValue(context, MemberNames._id.toString());
+	}
+
+	/**
+	 * Set value of _id
+	 * @param _id
+	 */
+	public final void set_id(java.lang.Long _id)
+	{
+		set_id(getContext(), _id);
+	}
+
+	/**
+	 * Set value of _id
+	 * @param context
+	 * @param _id
+	 */
+	public final void set_id(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Long _id)
+	{
+		getMendixObject().setValue(context, MemberNames._id.toString(), _id);
+	}
+
+	/**
+	 * @return value of Node_Scene
+	 */
+	public final scenegraph.proxies.Scene getNode_Scene() throws com.mendix.core.CoreException
+	{
+		return getNode_Scene(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of Node_Scene
+	 */
+	public final scenegraph.proxies.Scene getNode_Scene(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
+	{
+		scenegraph.proxies.Scene result = null;
+		com.mendix.systemwideinterfaces.core.IMendixIdentifier identifier = getMendixObject().getValue(context, MemberNames.Node_Scene.toString());
+		if (identifier != null)
+			result = scenegraph.proxies.Scene.load(context, identifier);
+		return result;
+	}
+
+	/**
+	 * Set value of Node_Scene
+	 * @param node_scene
+	 */
+	public final void setNode_Scene(scenegraph.proxies.Scene node_scene)
+	{
+		setNode_Scene(getContext(), node_scene);
+	}
+
+	/**
+	 * Set value of Node_Scene
+	 * @param context
+	 * @param node_scene
+	 */
+	public final void setNode_Scene(com.mendix.systemwideinterfaces.core.IContext context, scenegraph.proxies.Scene node_scene)
+	{
+		if (node_scene == null)
+			getMendixObject().setValue(context, MemberNames.Node_Scene.toString(), null);
+		else
+			getMendixObject().setValue(context, MemberNames.Node_Scene.toString(), node_scene.getMendixObject().getId());
+	}
+
+	/**
+	 * @return value of Node_Node
+	 */
+	public final scenegraph.proxies.Node getNode_Node() throws com.mendix.core.CoreException
+	{
+		return getNode_Node(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of Node_Node
+	 */
+	public final scenegraph.proxies.Node getNode_Node(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
+	{
+		scenegraph.proxies.Node result = null;
+		com.mendix.systemwideinterfaces.core.IMendixIdentifier identifier = getMendixObject().getValue(context, MemberNames.Node_Node.toString());
+		if (identifier != null)
+			result = scenegraph.proxies.Node.load(context, identifier);
+		return result;
+	}
+
+	/**
+	 * Set value of Node_Node
+	 * @param node_node
+	 */
+	public final void setNode_Node(scenegraph.proxies.Node node_node)
+	{
+		setNode_Node(getContext(), node_node);
+	}
+
+	/**
+	 * Set value of Node_Node
+	 * @param context
+	 * @param node_node
+	 */
+	public final void setNode_Node(com.mendix.systemwideinterfaces.core.IContext context, scenegraph.proxies.Node node_node)
+	{
+		if (node_node == null)
+			getMendixObject().setValue(context, MemberNames.Node_Node.toString(), null);
+		else
+			getMendixObject().setValue(context, MemberNames.Node_Node.toString(), node_node.getMendixObject().getId());
+	}
+
+	/**
+	 * @return the IMendixObject instance of this proxy for use in the Core interface.
+	 */
+	public final com.mendix.systemwideinterfaces.core.IMendixObject getMendixObject()
+	{
+		return nodeMendixObject;
+	}
+
+	/**
+	 * @return the IContext instance of this proxy, or null if no IContext instance was specified at initialization.
+	 */
+	public final com.mendix.systemwideinterfaces.core.IContext getContext()
+	{
+		return context;
+	}
+
+	@java.lang.Override
+	public boolean equals(Object obj)
+	{
+		if (obj == this)
+			return true;
+
+		if (obj != null && getClass().equals(obj.getClass()))
+		{
+			final scenegraph.proxies.Node that = (scenegraph.proxies.Node) obj;
+			return getMendixObject().equals(that.getMendixObject());
+		}
+		return false;
+	}
+
+	@java.lang.Override
+	public int hashCode()
+	{
+		return getMendixObject().hashCode();
+	}
+
+	/**
+	 * @return String name of this class
+	 */
+	public static java.lang.String getType()
+	{
+		return "SceneGraph.Node";
+	}
+
+	/**
+	 * @return String GUID from this object, format: ID_0000000000
+	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
+	 */
+	@java.lang.Deprecated
+	public java.lang.String getGUID()
+	{
+		return "ID_" + getMendixObject().getId().toLong();
+	}
+}
