+++ test/javasource/scenegraph/proxies/Scene.java
@@ -1,0 +1,332 @@
+// This file was generated by Mendix Modeler.
+//
+// WARNING: Code you write here will be lost the next time you deploy the project.
+
+package scenegraph.proxies;
+
+public class Scene
+{
+	private final com.mendix.systemwideinterfaces.core.IMendixObject sceneMendixObject;
+
+	private final com.mendix.systemwideinterfaces.core.IContext context;
+
+	/**
+	 * Internal name of this entity
+	 */
+	public static final java.lang.String entityName = "SceneGraph.Scene";
+
+	/**
+	 * Enum describing members of this entity
+	 */
+	public enum MemberNames
+	{
+		Name("Name"),
+		_Id("_Id"),
+		Selected("SceneGraph.Selected"),
+		Scene_Camera("SceneGraph.Scene_Camera");
+
+		private java.lang.String metaName;
+
+		MemberNames(java.lang.String s)
+		{
+			metaName = s;
+		}
+
+		@java.lang.Override
+		public java.lang.String toString()
+		{
+			return metaName;
+		}
+	}
+
+	public Scene(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		this(context, com.mendix.core.Core.instantiate(context, "SceneGraph.Scene"));
+	}
+
+	protected Scene(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject sceneMendixObject)
+	{
+		if (sceneMendixObject == null)
+			throw new java.lang.IllegalArgumentException("The given object cannot be null.");
+		if (!com.mendix.core.Core.isSubClassOf("SceneGraph.Scene", sceneMendixObject.getType()))
+			throw new java.lang.IllegalArgumentException("The given object is not a SceneGraph.Scene");
+
+		this.sceneMendixObject = sceneMendixObject;
+		this.context = context;
+	}
+
+	/**
+	 * @deprecated Use 'Scene.load(IContext, IMendixIdentifier)' instead.
+	 */
+	@java.lang.Deprecated
+	public static scenegraph.proxies.Scene initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		return scenegraph.proxies.Scene.load(context, mendixIdentifier);
+	}
+
+	/**
+	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
+	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
+	 */
+	public static scenegraph.proxies.Scene initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
+	{
+		return new scenegraph.proxies.Scene(context, mendixObject);
+	}
+
+	public static scenegraph.proxies.Scene load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
+		return scenegraph.proxies.Scene.initialize(context, mendixObject);
+	}
+
+	public static java.util.List<scenegraph.proxies.Scene> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
+	{
+		java.util.List<scenegraph.proxies.Scene> result = new java.util.ArrayList<scenegraph.proxies.Scene>();
+		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//SceneGraph.Scene" + xpathConstraint))
+			result.add(scenegraph.proxies.Scene.initialize(context, obj));
+		return result;
+	}
+
+	/**
+	 * Commit the changes made on this proxy object.
+	 */
+	public final void commit() throws com.mendix.core.CoreException
+	{
+		com.mendix.core.Core.commit(context, getMendixObject());
+	}
+
+	/**
+	 * Commit the changes made on this proxy object using the specified context.
+	 */
+	public final void commit(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
+	{
+		com.mendix.core.Core.commit(context, getMendixObject());
+	}
+
+	/**
+	 * Delete the object.
+	 */
+	public final void delete()
+	{
+		com.mendix.core.Core.delete(context, getMendixObject());
+	}
+
+	/**
+	 * Delete the object using the specified context.
+	 */
+	public final void delete(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		com.mendix.core.Core.delete(context, getMendixObject());
+	}
+	/**
+	 * @return value of Name
+	 */
+	public final java.lang.String getName()
+	{
+		return getName(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of Name
+	 */
+	public final java.lang.String getName(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		return (java.lang.String) getMendixObject().getValue(context, MemberNames.Name.toString());
+	}
+
+	/**
+	 * Set value of Name
+	 * @param name
+	 */
+	public final void setName(java.lang.String name)
+	{
+		setName(getContext(), name);
+	}
+
+	/**
+	 * Set value of Name
+	 * @param context
+	 * @param name
+	 */
+	public final void setName(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String name)
+	{
+		getMendixObject().setValue(context, MemberNames.Name.toString(), name);
+	}
+
+	/**
+	 * @return value of _Id
+	 */
+	public final java.lang.Long get_Id()
+	{
+		return get_Id(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of _Id
+	 */
+	public final java.lang.Long get_Id(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		return (java.lang.Long) getMendixObject().getValue(context, MemberNames._Id.toString());
+	}
+
+	/**
+	 * Set value of _Id
+	 * @param _id
+	 */
+	public final void set_Id(java.lang.Long _id)
+	{
+		set_Id(getContext(), _id);
+	}
+
+	/**
+	 * Set value of _Id
+	 * @param context
+	 * @param _id
+	 */
+	public final void set_Id(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Long _id)
+	{
+		getMendixObject().setValue(context, MemberNames._Id.toString(), _id);
+	}
+
+	/**
+	 * @return value of Selected
+	 */
+	public final scenegraph.proxies.Primitive getSelected() throws com.mendix.core.CoreException
+	{
+		return getSelected(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of Selected
+	 */
+	public final scenegraph.proxies.Primitive getSelected(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
+	{
+		scenegraph.proxies.Primitive result = null;
+		com.mendix.systemwideinterfaces.core.IMendixIdentifier identifier = getMendixObject().getValue(context, MemberNames.Selected.toString());
+		if (identifier != null)
+			result = scenegraph.proxies.Primitive.load(context, identifier);
+		return result;
+	}
+
+	/**
+	 * Set value of Selected
+	 * @param selected
+	 */
+	public final void setSelected(scenegraph.proxies.Primitive selected)
+	{
+		setSelected(getContext(), selected);
+	}
+
+	/**
+	 * Set value of Selected
+	 * @param context
+	 * @param selected
+	 */
+	public final void setSelected(com.mendix.systemwideinterfaces.core.IContext context, scenegraph.proxies.Primitive selected)
+	{
+		if (selected == null)
+			getMendixObject().setValue(context, MemberNames.Selected.toString(), null);
+		else
+			getMendixObject().setValue(context, MemberNames.Selected.toString(), selected.getMendixObject().getId());
+	}
+
+	/**
+	 * @return value of Scene_Camera
+	 */
+	public final scenegraph.proxies.Camera getScene_Camera() throws com.mendix.core.CoreException
+	{
+		return getScene_Camera(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of Scene_Camera
+	 */
+	public final scenegraph.proxies.Camera getScene_Camera(com.mendix.systemwideinterfaces.core.IContext context) throws com.mendix.core.CoreException
+	{
+		scenegraph.proxies.Camera result = null;
+		com.mendix.systemwideinterfaces.core.IMendixIdentifier identifier = getMendixObject().getValue(context, MemberNames.Scene_Camera.toString());
+		if (identifier != null)
+			result = scenegraph.proxies.Camera.load(context, identifier);
+		return result;
+	}
+
+	/**
+	 * Set value of Scene_Camera
+	 * @param scene_camera
+	 */
+	public final void setScene_Camera(scenegraph.proxies.Camera scene_camera)
+	{
+		setScene_Camera(getContext(), scene_camera);
+	}
+
+	/**
+	 * Set value of Scene_Camera
+	 * @param context
+	 * @param scene_camera
+	 */
+	public final void setScene_Camera(com.mendix.systemwideinterfaces.core.IContext context, scenegraph.proxies.Camera scene_camera)
+	{
+		if (scene_camera == null)
+			getMendixObject().setValue(context, MemberNames.Scene_Camera.toString(), null);
+		else
+			getMendixObject().setValue(context, MemberNames.Scene_Camera.toString(), scene_camera.getMendixObject().getId());
+	}
+
+	/**
+	 * @return the IMendixObject instance of this proxy for use in the Core interface.
+	 */
+	public final com.mendix.systemwideinterfaces.core.IMendixObject getMendixObject()
+	{
+		return sceneMendixObject;
+	}
+
+	/**
+	 * @return the IContext instance of this proxy, or null if no IContext instance was specified at initialization.
+	 */
+	public final com.mendix.systemwideinterfaces.core.IContext getContext()
+	{
+		return context;
+	}
+
+	@java.lang.Override
+	public boolean equals(Object obj)
+	{
+		if (obj == this)
+			return true;
+
+		if (obj != null && getClass().equals(obj.getClass()))
+		{
+			final scenegraph.proxies.Scene that = (scenegraph.proxies.Scene) obj;
+			return getMendixObject().equals(that.getMendixObject());
+		}
+		return false;
+	}
+
+	@java.lang.Override
+	public int hashCode()
+	{
+		return getMendixObject().hashCode();
+	}
+
+	/**
+	 * @return String name of this class
+	 */
+	public static java.lang.String getType()
+	{
+		return "SceneGraph.Scene";
+	}
+
+	/**
+	 * @return String GUID from this object, format: ID_0000000000
+	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
+	 */
+	@java.lang.Deprecated
+	public java.lang.String getGUID()
+	{
+		return "ID_" + getMendixObject().getId().toLong();
+	}
+}
