+++ test/javasource/scenegraph/proxies/Text.java
@@ -1,0 +1,202 @@
+// This file was generated by Mendix Modeler.
+//
+// WARNING: Code you write here will be lost the next time you deploy the project.
+
+package scenegraph.proxies;
+
+public class Text extends scenegraph.proxies.Primitive
+{
+	/**
+	 * Internal name of this entity
+	 */
+	public static final java.lang.String entityName = "SceneGraph.Text";
+
+	/**
+	 * Enum describing members of this entity
+	 */
+	public enum MemberNames
+	{
+		val("val"),
+		sz("sz"),
+		Name("Name"),
+		x("x"),
+		y("y"),
+		z("z"),
+		rotx("rotx"),
+		roty("roty"),
+		rotz("rotz"),
+		visible("visible"),
+		color("color"),
+		Primitive_Node("SceneGraph.Primitive_Node"),
+		Selected("SceneGraph.Selected"),
+		Texture("SceneGraph.Texture");
+
+		private java.lang.String metaName;
+
+		MemberNames(java.lang.String s)
+		{
+			metaName = s;
+		}
+
+		@java.lang.Override
+		public java.lang.String toString()
+		{
+			return metaName;
+		}
+	}
+
+	public Text(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		this(context, com.mendix.core.Core.instantiate(context, "SceneGraph.Text"));
+	}
+
+	protected Text(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject textMendixObject)
+	{
+		super(context, textMendixObject);
+		if (!com.mendix.core.Core.isSubClassOf("SceneGraph.Text", textMendixObject.getType()))
+			throw new java.lang.IllegalArgumentException("The given object is not a SceneGraph.Text");
+	}
+
+	/**
+	 * @deprecated Use 'Text.load(IContext, IMendixIdentifier)' instead.
+	 */
+	@java.lang.Deprecated
+	public static scenegraph.proxies.Text initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		return scenegraph.proxies.Text.load(context, mendixIdentifier);
+	}
+
+	/**
+	 * Initialize a proxy using context (recommended). This context will be used for security checking when the get- and set-methods without context parameters are called.
+	 * The get- and set-methods with context parameter should be used when for instance sudo access is necessary (IContext.createSudoClone() can be used to obtain sudo access).
+	 */
+	public static scenegraph.proxies.Text initialize(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixObject mendixObject)
+	{
+		return new scenegraph.proxies.Text(context, mendixObject);
+	}
+
+	public static scenegraph.proxies.Text load(com.mendix.systemwideinterfaces.core.IContext context, com.mendix.systemwideinterfaces.core.IMendixIdentifier mendixIdentifier) throws com.mendix.core.CoreException
+	{
+		com.mendix.systemwideinterfaces.core.IMendixObject mendixObject = com.mendix.core.Core.retrieveId(context, mendixIdentifier);
+		return scenegraph.proxies.Text.initialize(context, mendixObject);
+	}
+
+	public static java.util.List<scenegraph.proxies.Text> load(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String xpathConstraint) throws com.mendix.core.CoreException
+	{
+		java.util.List<scenegraph.proxies.Text> result = new java.util.ArrayList<scenegraph.proxies.Text>();
+		for (com.mendix.systemwideinterfaces.core.IMendixObject obj : com.mendix.core.Core.retrieveXPathQuery(context, "//SceneGraph.Text" + xpathConstraint))
+			result.add(scenegraph.proxies.Text.initialize(context, obj));
+		return result;
+	}
+
+	/**
+	 * @return value of val
+	 */
+	public final java.lang.String getval()
+	{
+		return getval(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of val
+	 */
+	public final java.lang.String getval(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		return (java.lang.String) getMendixObject().getValue(context, MemberNames.val.toString());
+	}
+
+	/**
+	 * Set value of val
+	 * @param val
+	 */
+	public final void setval(java.lang.String val)
+	{
+		setval(getContext(), val);
+	}
+
+	/**
+	 * Set value of val
+	 * @param context
+	 * @param val
+	 */
+	public final void setval(com.mendix.systemwideinterfaces.core.IContext context, java.lang.String val)
+	{
+		getMendixObject().setValue(context, MemberNames.val.toString(), val);
+	}
+
+	/**
+	 * @return value of sz
+	 */
+	public final java.lang.Double getsz()
+	{
+		return getsz(getContext());
+	}
+
+	/**
+	 * @param context
+	 * @return value of sz
+	 */
+	public final java.lang.Double getsz(com.mendix.systemwideinterfaces.core.IContext context)
+	{
+		return (java.lang.Double) getMendixObject().getValue(context, MemberNames.sz.toString());
+	}
+
+	/**
+	 * Set value of sz
+	 * @param sz
+	 */
+	public final void setsz(java.lang.Double sz)
+	{
+		setsz(getContext(), sz);
+	}
+
+	/**
+	 * Set value of sz
+	 * @param context
+	 * @param sz
+	 */
+	public final void setsz(com.mendix.systemwideinterfaces.core.IContext context, java.lang.Double sz)
+	{
+		getMendixObject().setValue(context, MemberNames.sz.toString(), sz);
+	}
+
+	@java.lang.Override
+	public boolean equals(Object obj)
+	{
+		if (obj == this)
+			return true;
+
+		if (obj != null && getClass().equals(obj.getClass()))
+		{
+			final scenegraph.proxies.Text that = (scenegraph.proxies.Text) obj;
+			return getMendixObject().equals(that.getMendixObject());
+		}
+		return false;
+	}
+
+	@java.lang.Override
+	public int hashCode()
+	{
+		return getMendixObject().hashCode();
+	}
+
+	/**
+	 * @return String name of this class
+	 */
+	public static java.lang.String getType()
+	{
+		return "SceneGraph.Text";
+	}
+
+	/**
+	 * @return String GUID from this object, format: ID_0000000000
+	 * @deprecated Use getMendixObject().getId().toLong() to get a unique identifier for this object.
+	 */
+	@java.lang.Override
+	@java.lang.Deprecated
+	public java.lang.String getGUID()
+	{
+		return "ID_" + getMendixObject().getId().toLong();
+	}
+}
