+++ test/javasource/scenegraph/proxies/constants/Constants.java
@@ -1,0 +1,17 @@
+// This file was generated by Mendix Modeler 7.23.
+//
+// WARNING: Code you write here will be lost the next time you deploy the project.
+
+package scenegraph.proxies.constants;
+
+import com.mendix.core.Core;
+
+public class Constants
+{
+	// These are the constants for the SceneGraph module
+
+	public static java.lang.Double getPI()
+	{
+		return (java.lang.Double)Core.getConfiguration().getConstantValue("SceneGraph.PI");
+	}
+}
